LayoutTests/crypto/webkitSubtle/rsa-export-generated-keys-expected.txt

CONSOLE MESSAGE: line 26: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test exporting a generated RSA keypair with hash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Generating RSA-PKCS1-v1.5 keyPair...
PASS crypto.webkitSubtle.exportKey(null, key) threw exception TypeError: Unknown key format.
PASS crypto.webkitSubtle.exportKey(undefined, key) threw exception TypeError: Unknown key format.
PASS crypto.webkitSubtle.exportKey({}, key) threw exception TypeError: Unknown key format.
PASS crypto.webkitSubtle.exportKey("", key) threw exception TypeError: Unknown key format.
PASS crypto.webkitSubtle.exportKey("foobar", key) threw exception TypeError: Unknown key format.
PASS key.publicKey.algorithm.hash.name is defined.
PASS key.privateKey.algorithm.hash.name is defined.

Exporting public key as JWK...
PASS exportedJWK.kty is 'RSA'
PASS exportedJWK.alg is 'RS256'
PASS exportedJWK.ext is true
PASS exportedJWK.use is undefined
PASS exportedJWK.key_ops is ['sign', 'verify']

Exporting private key as JWK...
PASS exportedJWK.kty is 'RSA'
PASS exportedJWK.alg is 'RS256'
PASS exportedJWK.ext is true
PASS exportedJWK.use is undefined
PASS exportedJWK.key_ops is ['sign', 'verify']
PASS successfullyParsed is true

TEST COMPLETE
